+++ 10-05-2017/p3.pas
@@ -1,5 +1,5 @@
 VAR
-    x1, y1, x2, y2 : INTEGER;
+    x1, y1, x2, y2 : REAL;
 BEGIN
     readln(x1, y1, x2, y2);
     writeln('l: ', sqrt(sqr(x2-x1) + sqr(y2-y1)));
