+++ 10-05-2017/p4.pas
@@ -1,6 +1,6 @@
 VAR
     x1, y1, x2, y2, x3, y3,
-    a, b, c, s : INTEGER;
+    a, b, c, s : REAL;
 BEGIN
     readln(x1, y1, x2, y2, x3, y3);
     a := sqrt(sqr(x2-x1) + sqr(y2-y1));
