+++ 10-05-2017/p5.pas
@@ -1,7 +1,7 @@
 USES math;
 
 VAR
-    a, b, al : INTEGER;
+    a, b, al : REAL;
 BEGIN
     read(a, b, al);
     writeln('S: ', tan(al)*(a-b)/2*(a + b)/2);
